--- U-Net/U-Net/2024-06-26/Model_Data/history.json
@@ -1,1 +1,0 @@
-{"accuracy": [0.44311121106147766], "loss": [43.86216354370117], "val_accuracy": [0.9586508274078369], "val_loss": [39.879451751708984]}
